changed CAN signal mappings
These changes result in the ability to eliminate the need to have a connection to the OBD connector. Most of the changes were on the gateway script.

From OP perspective it is still receiving the data from CAN1 of the Panda.

using 0x2C1 for GAS_PEDAL on Driving bus instead of 0x49B on OBD6/14